requirements.txt
atomicwrites==1.1.5
attrdict==2.0.0
attrs==18.1.0
bitcoin==1.1.42
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cytoolz==0.9.0.1
eth-abi==1.1.1
eth-account==0.2.3
eth-hash==0.1.4
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-tester==0.1.0b26
eth-testrpc==1.3.5
eth-utils==1.0.3
ethereum==1.6.1
hexbytes==0.1.0
idna==2.6
json-rpc==1.11.0
lru-dict==1.1.6
more-itertools==4.2.0
parsimonious==0.8.0
pbkdf2==1.3
pluggy==0.6.0
py==1.5.3
py-solc==3.0.0
pycparser==2.18
pycryptodome==3.6.1
pyethash==0.1.27
pysha3==1.0.2
pytest==3.6.0
PyYAML==3.12
repoze.lru==0.7
requests==2.18.4
rlp==0.6.0
scrypt==0.8.6
secp256k1==0.13.2
semantic-version==2.6.0
six==1.11.0
toolz==0.9.0
urllib3==1.22
web3==4.2.1
websockets==5.0.1
Werkzeug==0.14.1
